+++ grelmicro/log/__init__.py
@@ -8,11 +8,7 @@ from grelmicro._config import resolve_config
 from grelmicro.log._apply import apply as _apply
 from grelmicro.log._component import Log
 from grelmicro.log._dedup import DuplicateFilter, DuplicateFilterConfig
-from grelmicro.log._probes import (
-    DEFAULT_PROBE_PATHS,
-    ProbeFilter,
-    silence_probe_access_logs,
-)
+from grelmicro.log._probes import ProbeFilter
 from grelmicro.log._ratelimit import (
     RateLimitFilter,
     RateLimitFilterConfig,
@@ -142,7 +138,6 @@ def configure_with(
 
 
 __all__ = [
-    "DEFAULT_PROBE_PATHS",
     "DuplicateFilter",
     "DuplicateFilterConfig",
     "ErrorDict",
@@ -156,5 +151,4 @@ __all__ = [
     "RateLimitFilterConfig",
     "configure",
     "configure_with",
-    "silence_probe_access_logs",
 ]